Renamed AppointmentListItemVaos to AppointmentListItem and deleted old AppointmentListItem.
